requirements.txt
django==5.0.2
djangorestframework 
